Finished the usage of red cubes, rewrote the potential system nearly fully now.
Created a handler for UseCashItem. So far this only fully handles the
red and bonus cubes. Simple cubes (revolutionary, super, etc...) are
working but have no values for tier up/3rd line chance. Harder cubes to
implement will be black, violet and memory cubes.

Updated ServerHandler to have an option for a spam filter, and to
actually show the stack trace of an exception. Also updated UseMagnify
to reflect the changes made to the potential system.

Lastly, this time actually 'fixed' the ArrayIndexOutOfBoundsException
given by CharacterWithSecondPassword. It's skips a single byte that's
been causing trouble.