Null check for Element elem.
Null gets passed through at CalculateMaxMagicDamagePerHit but that same null value does not get handled properly at ElementalStaffAttackBonus. When getElementBoost() is unable to find any elem inside elemBoosts, it returns 0, so I just removed it if elem is found to be null.

Prior to this, skills such as Three Snails were displaying damage but failed to be parsed. Now they work as intended.