Fix for selling items.
The check was failing, causing all items to be sold for 1 Meso. Removed the check and I was able to sell items for the normal price.

Have not tested throwing stars or bullets, just regular items.
For some reason MapleStory hard crashes when purchasing items, will look into it further.