Almost everything created for the star force system
Enhancing works, although a boom will simply make the item go back to
0*, instead of making it an item trace. Additionally, it currently uses
a var in Equip to track what stats it should remove if a star fails,
thus failing if the server restarts or something.

Created a class to hold all constants regarding items, has to be
populated by 'old' methods in GameConstants.

Made a method in MapleCharacter to check and add meso, should be always
used when removing mesos from a character.

Lastly, added some GM commands that could be useful for testing star
force.